# niki/target/niki/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "niki"
    },
    "host": "",
    "basePath": "/niki/rest",
    "schemes": [
        "http"
    ],
    "paths": {
        "/cities": {
            "get": {
                "consumes": [
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "max",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "start",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/City"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/City"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "headers": {
                            "Location": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/cities/{id}": {
            "get": {
                "consumes": [
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "$ref": "#/definitions/City"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "headers": {
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/City"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "headers": {
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "headers": {
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "headers": {
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "headers": {
                        },
                        "schema": {
                            "$ref": "#/definitions/City"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                ],
                "produces": [
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "headers": {
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "headers": {
                        }
                    }
                }
            }
        },
        "/companies": {
            "get": {
                "consumes": [
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "max",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "start",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Company"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Company"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "headers": {
                            "Location": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/companies/{id}": {
            "get": {
                "consumes": [
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "$ref": "#/definitions/Company"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "headers": {
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Company"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "headers": {
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "headers": {
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "headers": {
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "headers": {
                        },
                        "schema": {
                            "$ref": "#/definitions/Company"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                ],
                "produces": [
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "headers": {
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "headers": {
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "City": {
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "province": {
                    "$ref": "#/definitions/Province"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "Company": {
            "properties": {
                "cellphone": {
                    "type": "string"
                },
                "city": {
                    "$ref": "#/definitions/City"
                },
                "complement": {
                    "type": "string"
                },
                "country": {
                    "$ref": "#/definitions/Country"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "employees": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Employee"
                    }
                },
                "houseNumber": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "landline": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "neighborhood": {
                    "$ref": "#/definitions/Neighborhood"
                },
                "postalCode": {
                    "type": "string"
                },
                "province": {
                    "$ref": "#/definitions/Province"
                },
                "street": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "Country": {
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "Employee": {
            "properties": {
                "admission": {
                    "type": "object"
                },
                "birth": {
                    "type": "object"
                },
                "cellphone": {
                    "type": "string"
                },
                "city": {
                    "$ref": "#/definitions/City"
                },
                "company": {
                    "$ref": "#/definitions/Company"
                },
                "complement": {
                    "type": "string"
                },
                "country": {
                    "$ref": "#/definitions/Country"
                },
                "dealEnd": {
                    "type": "object"
                },
                "dealStart": {
                    "type": "object"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "Female",
                        "Male"
                    ]
                },
                "houseNumber": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "identityIssuer": {
                    "type": "string"
                },
                "identityNumber": {
                    "type": "string"
                },
                "landline": {
                    "type": "string"
                },
                "legacySheetId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "neighborhood": {
                    "$ref": "#/definitions/Neighborhood"
                },
                "position": {
                    "$ref": "#/definitions/Position"
                },
                "postalCode": {
                    "type": "string"
                },
                "province": {
                    "$ref": "#/definitions/Province"
                },
                "schedule": {
                    "$ref": "#/definitions/Schedule"
                },
                "socialSecurityId": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "taxPayerNumber": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
        "Neighborhood": {
            "properties": {
                "city": {
                    "$ref": "#/definitions/City"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "Position": {
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "Province": {
            "properties": {
                "code": {
                    "type": "string"
                },
                "country": {
                    "$ref": "#/definitions/Country"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "Schedule": {
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        }
    }
}